Type system cleanup
