+++ ical/event.py
@@ -5,6 +5,7 @@
 from __future__ import annotations
 
 import datetime
+import enum
 import logging
 from typing import Any, Optional, Union
 
@@ -13,11 +14,18 @@ from pydantic import Field, root_validator
 from ._types import ComponentModel, validate_until_dtstart
 from .alarm import Alarm
 from .parsing.property import ParsedProperty
-from .types import CalAddress, Geo, Priority, Recur, RequestStatus, Uri
-from .types.const import Classification, EventStatus
+from .types import CalAddress, Classification, Geo, Priority, Recur, RequestStatus, Uri
 from .util import dtstamp_factory, normalize_datetime, uid_factory
 
 _LOGGER = logging.getLogger(__name__)
+
+
+class EventStatus(str, enum.Enum):
+    """Status or confirmation of the event."""
+
+    CONFIRMED = "CONFIRMED"
+    TENTATIVE = "TENTATIVE"
+    CANCELLED = "CANCELLED"
 
 
 class Event(ComponentModel):
