+++ ical/todo.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import datetime
+import enum
 from typing import Any, Optional, Union
 
 from pydantic import Field, root_validator
@@ -10,9 +11,17 @@ from pydantic import Field, root_validator
 from ._types import ComponentModel, validate_until_dtstart
 from .alarm import Alarm
 from .parsing.property import ParsedProperty
-from .types import CalAddress, Geo, Priority, Recur, RequestStatus, Uri
-from .types.const import Classification, TodoStatus
+from .types import CalAddress, Classification, Geo, Priority, Recur, RequestStatus, Uri
 from .util import dtstamp_factory, normalize_datetime, uid_factory
+
+
+class TodoStatus(str, enum.Enum):
+    """Status or confirmation of the to-do."""
+
+    NEEDS_ACTION = "NEEDS-ACTION"
+    COMPLETED = "COMPLETED"
+    IN_PROCESS = "IN-PROCESS"
+    CANCELLED = "CANCELLED"
 
 
 class Todo(ComponentModel):
