+++ ical/types/__init__.py
@@ -2,21 +2,23 @@
 
 # Import all types for the registry
 from . import integer  # noqa: F401
-from . import recur  # noqa: F401
 from . import boolean, date, date_time, duration  # noqa: F401
 from . import float as float_pkg  # noqa: F401
 from .cal_address import CalAddress
+from .const import Classification
 from .geo import Geo
-from .period import Period
+from .period import FreeBusyType, Period
 from .priority import Priority
-from .recur import Recur
+from .recur import Frequency, Recur, Weekday, WeekdayValue
 from .request_status import RequestStatus
 from .uri import Uri
 from .utc_offset import UtcOffset
 
 __all__ = [
-    "const",
     "CalAddress",
+    "Classification",
+    "Frequency",
+    "FreeBusyType",
     "Geo",
     "Period",
     "Priority",
@@ -24,4 +26,6 @@ __all__ = [
     "RequestStatus",
     "UtcOffset",
     "Uri",
+    "Weekday",
+    "WeekdayValue",
 ]
