+++ tests/types/test_period.py
@@ -5,10 +5,11 @@ import datetime
 import pytest
 from pydantic import ValidationError
 
-from ical._types import ICS_ENCODERS, ComponentModel
+from ical._types import ComponentModel
 from ical.parsing.component import ParsedComponent
 from ical.parsing.property import ParsedProperty
 from ical.types import Period
+from ical.types.data_types import DATA_TYPE
 
 
 class FakeModel(ComponentModel):
@@ -19,7 +20,7 @@ class FakeModel(ComponentModel):
     class Config:
         """Pydantic model configuration."""
 
-        json_encoders = ICS_ENCODERS
+        json_encoders = DATA_TYPE.encode_property_json
 
 
 def test_period() -> None:
